Dashboard Profile loading bugs has been fixed by sourabh and AuthProvider updated
